added controllers in the application with .env file
