+++ .env.sample
@@ -1,0 +1,2 @@
+PORT = 3000
+DATABASE_URL = PUT DATABASE_URL here
