+++ src/app.ts
@@ -1,13 +1,15 @@
 import express, { Application, Request, Response } from "express";
 import userRouter from "./routes/users";
+import * as dotenv from "dotenv";
 import mongoose from "mongoose";
 
+dotenv.config();
 const app: Application = express();
 
-const PORT: number = 4000;
+const PORT: any = process.env.PORT;
 
 app.get("/", (req: Request, res: Response) => {
-  res.send("Hi this is our first page");
+  res.send("hello world");
 });
 
 app.use(express.json());
