+++ src/routes/users.ts
@@ -1,35 +1,21 @@
 import express, { Router, Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";
-import bcrypt from "bcrypt";
+import { getUsers, createUser } from "../controllers/users";
+
 const router: Router = express.Router();
 
 let password: (x: string) => string = function (x) {
   return x;
 };
 
-interface Users {
-  id: string;
-  username: string;
-  email: string;
-  password: any;
-  name: string;
-}
-
 let users: any = [];
 
-router.get("/", (req: Request, res: Response) => {
-  res.status(200).json(users);
-});
+router.get("/", getUsers);
 
-router.post("/", (req: Request, res: Response) => {
-  let newUser: Users = {
-    id: uuidv4().toString(),
-    username: req.body.username,
-    email: req.body.email,
-    password: bcrypt.hash(req.body.password, 10, (err, hash) => hash),
-    name: req.body.name,
-  };
-  users.push(newUser);
+router.post("/", createUser);
+
+router.patch("/:id", (req: Request, res: Response) => {
+  res.send(users.filter((user: any) => user.id === req.params.id));
 });
 
 export default router;
